Evil trick to re-initialize
Handle with care. jQuery does not allow to re-init the page. This
commit will remove the current js.php from the DOM and add it again so
that jQuery gets initialised again (leaving us with yet another jQuery
instance due to noConflict() - which is OKish). It is very evil and
might lead to problems that I’m not yet aware of. But it works.